requirements.txt
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cssselect2==0.8.0
defusedxml==0.7.1
idna==3.10
pillow==11.2.1
pycparser==2.22
pyoxipng==9.1.0
requests==2.32.3
tinycss2==1.4.0
urllib3==2.4.0
webencodings==0.5.1
